Create SelectedText struct
This struct is going to hold the data of what text is currently selected, the
decision of whether text should be highlighted should be based on the current
mode for now. There are arguments that maybe passing that struct around could be
useful in certain situations outside of visual mode, such as when searching
text, for example.