Added highlighting in visual mode
Created the draw_highlighted_row function that's called whenever any part of a
row is highlighted.
There are some caveats:
- Currently there are 2 checks for whether hl_text.start is greater than
hl_text.end, one in draw_rows, and another in draw_row. I'm still not sure how
to fix that, but for now it works.
- Maybe the SelectedText struct should be made as to always guarantee that
hl_text.start <= hl_text.end. But that would mean that the cursor may correspond
do hl_text.start or hl_text.end depending on context, which would just be a pain
to deal with. I think I prefer having start always be stationary and let end do
its thing, but it's still something to keep in mind.
- The draw_highlighted_rows function does not work when hl_text.end is directly
on a newline, since it tries to add 1 to x. I haven't tackled that yet since I
wanted to first create a basic function that was working so I could move on to
other things, but it should be a trivial fix (albeit a bit verbose, mayba)